pyproject.toml
[project]
name = "tower-of-hanoi-experiment"
version = "0.1.0"
description = "Replicate Tower of Hanoi experiments with Claude API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.18.0",
    "click>=8.1.7",
    "tqdm>=4.66.1",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"